requirements.txt
jupyter~=1.0.0
pip~=24.2
attrs~=23.1.0
distro~=1.9.0
wheel~=0.41.0
retry~=0.9.2
PySocks~=1.7.1
protobuf~=4.22.3
cryptography~=43.0.1
Pillow~=10.1.0
asyncio~=3.4.3
tornado~=6.3.2
Jinja2~=3.1.2
filelock~=3.13.1
redis~=4.5.4
setuptools~=50.3.2
matplotlib~=3.7.1
typing_extensions~=4.12.2
packaging~=24.1
trove-classifiers~=2024.10.16
zipp~=3.17.0
build~=1.2.2.post1
platformdirs~=4.3.6
crewai~=0.76.2
python-dotenv~=1.0.1
hpack~=4.0.0
hyperframe~=6.0.1
h2~=4.1.0
py~=1.11.0
pytest~=8.3.3
colorama~=0.4.6
uv~=0.4.26
bleach~=6.1.0
lxml~=4.9.2
beautifulsoup4~=4.12.3
soupsieve~=2.4
gym~=0.26.2
numpy~=1.26.4
importlib-metadata~=6.8.0
opencv-python~=4.9.0.80
gym-notices~=0.0.8
h11~=0.14.0
rsa~=4.9
pyasn1~=0.5.0
absl-py~=1.4.0
cffi~=1.17.1
gast~=0.4.0
grpcio~=1.67.0
grpcio-tools~=1.62.3
h5py~=3.12.1
ipython~=8.14.0
parso~=0.8.3
jedi~=0.18.2
Mako~=1.2.4
Babel~=2.14.0
Pygments~=2.15.1
MarkupSafe~=2.1.2
ollama~=0.3.3
openai~=1.52.2
pydantic~=2.9.2
litellm~=1.50.4
boto3~=1.35.42
httpx~=0.27.2
pytz~=2024.2
posthog~=3.7.0
langchain-community~=0.3.3
rank-bm25~=0.2.2
google-cloud-aiplatform~=1.70.0
chromadb~=0.4.24
qdrant-client~=1.12.0
rich~=13.9.2
markdown-it-py~=3.0.0
slack-sdk~=3.21.2
requests~=2.31.0
keras~=3.6.0
ipywidgets~=8.0.7
notebook~=7.0.7
trio~=0.22.0
outcome~=1.2.0
async-generator~=1.10
sniffio~=1.3.0
exceptiongroup~=1.1.1
sortedcontainers~=2.4.0
idna~=3.4
PyYAML~=6.0
multidict~=6.0.4
propcache~=0.2.0
auth0-python~=4.7.2
PyJWT~=2.9.0
aiohttp~=3.10.10
yarl~=1.15.4
pyproject_hooks~=1.2.0
libclang~=16.0.6
click~=8.1.7
httpcore~=1.0.6
certifi~=2024.8.30
mock~=4.0.3
jiter~=0.4.2
optree~=0.13.0
tensorflow~=2.18.0
ml-dtypes~=0.4.1
psutil~=5.9.5
huggingface-hub~=0.25.2
scipy~=1.12.0
torch~=2.0.0
pandas~=1.5.3
namex~=0.0.8
pyarrow~=17.0.0
tqdm~=4.65.0
schema~=0.7.7
mdurl~=0.1.2
overrides~=7.6.0
charset-normalizer~=3.1.0
six~=1.16.0
google-pasta~=0.2.0
proto-plus~=1.24.0
pypdf~=5.0.1
pysbd~=0.3.4
networkx~=3.1
selenium~=4.25.0
jsonpickle~=3.3.0
regex~=2024.9.11
decorator~=5.1.1
sympy~=1.11.1
mpmath~=1.3.0
typer~=0.9.0
shellingham~=1.5.4
agents~=1.4.0
httpx-sse~=0.4.0
pydantic_core~=2.23.4
fastavro~=1.9.7
tokenizers~=0.20.1
cohere~=5.11.0
botocore~=1.35.42
sagemaker~=2.232.2
tomli_w~=1.1.0
pyvis~=0.3.2
json_repair~=0.25.3
appdirs~=1.4.4
urllib3~=2.1.0
docker~=7.1.0
fsspec~=2024.9.0
pyasn1-modules~=0.3.0
cachetools~=5.3.0
google-api-python-client~=2.84.0
google-crc32c~=1.5.0
shapely~=2.0.6
python-dateutil~=2.8.2
joblib~=1.3.2
fastapi~=0.115.2
uvicorn~=0.32.0
starlette~=0.40.0
sklearn~=0.0.post4
scikit-learn~=1.4.0
tensorboard~=2.18.0
mlflow~=2.17.0
mlflow-skinny~=2.17.0
Werkzeug~=2.3.6
grpcio-status~=1.54.0
anyio~=3.6.2
orjson~=3.10.7
SQLAlchemy~=2.0.36
PyPika~=0.48.9
pytube~=15.0.0
uvloop~=0.21.0
wcwidth~=0.2.6
pluggy~=1.5.0
Deprecated~=1.2.14
iniconfig~=2.0.0
pexpect~=4.8.0
async-timeout~=4.0.2
gunicorn~=23.0.0
aiosignal~=1.3.1
frozenlist~=1.3.3
aiodns~=3.0.0
aiohappyeyeballs~=2.4.3
alembic~=1.13.3
importlib_resources~=6.4.5
asgiref~=3.5.0
backoff~=2.2.1
websockets~=13.1
traitlets~=5.9.0
stack-data~=0.6.2
matplotlib-inline~=0.1.6
pickleshare~=0.7.5
jupyter_client~=8.3.0
prompt-toolkit~=3.0.39
lancedb~=0.5.7
arrow~=1.3.0
deprecation~=2.1.0
instructor~=1.3.3
transformers~=4.28.1
tiktoken~=0.7.0
ptyprocess~=0.7.0
monotonic~=1.6
tzdata~=2023.3
httptools~=0.6.4
wsproto~=1.2.0
watchfiles~=0.24.0
tenacity~=8.5.0
onnxruntime~=1.19.2
kubernetes~=31.0.0
mmh3~=5.0.1
chroma-hnswlib~=0.7.3
docx2txt~=0.8
gptcache~=0.1.44
langchain~=0.3.4
faiss-cpu~=1.7.3
torchaudio~=2.0.1
Markdown~=3.4.4
oauthlib~=3.2.2
blinker~=1.6.2
annotated-types~=0.6.0
trio-websocket~=0.10.2
websocket-client~=1.8.0
langchain-core~=0.3.12
asttokens~=2.2.1
executing~=1.2.0
langsmith~=0.1.135
langchain-text-splitters~=0.3.0
elasticsearch~=8.7.0
requests-toolbelt~=1.0.0
pure-eval~=0.2.2
pycparser~=2.21
itsdangerous~=2.1.2
termcolor~=2.3.0
nltk~=3.9.1
wrapt~=1.15.0
embedchain~=0.1.123
langchain-cohere~=0.3.1
langchain-openai~=0.2.3
Flask~=2.3.2
google-auth-oauthlib~=1.0.0
pinecone-client~=2.2.1
docstring_parser~=0.16
jsonref~=1.1.0
referencing~=0.32.1
jsonschema~=4.23.0
webcolors~=1.13
jsonpointer~=2.4
uri-template~=1.3.0
isoduration~=20.11.0
fqdn~=1.5.1
rfc3339-validator~=0.1.4
jsonschema-specifications~=2023.12.1
durationpy~=0.9
requests-oauthlib~=1.3.1
opt-einsum~=3.3.0
flatbuffers~=24.3.25
astunparse~=1.6.3
dill~=0.3.9
tblib~=3.0.0
cloudpickle~=2.2.1
humanfriendly~=10.0
coloredlogs~=15.0.1
marshmallow~=3.22.0
bcrypt~=4.0.1
parameterized~=0.9.0
portalocker~=2.10.1
jsonpatch~=1.33
typing-inspect~=0.9.0
dataclasses-json~=0.6.7
langchain-experimental~=0.3.2
pydantic-settings~=2.6.0
mypy-extensions~=1.0.0
databricks-sdk~=0.34.0
PyGithub~=2.1.1
pymongo~=4.3.3